Added onReserve callback to the lobby view
git-svn-id: file:///srv/svn/repos/futurice_reservator@102 967a2e56-9f69-44db-a86b-3acc15697453

+++ android/Reservator/src/com/futurice/android/reservator/view/CalendarView.java
@@ -136,7 +136,7 @@ public class CalendarView extends LinearLayout implements OnClickListener {
 		CalendarMarker v = new CalendarMarker(this.getContext());
 		LinearLayout.LayoutParams lp = new LinearLayout.LayoutParams(
 				LayoutParams.MATCH_PARENT, 1,
-				(float) (endTime.getTimeInMillis() - begin.getTimeInMillis()));
+				(endTime.getTimeInMillis() - begin.getTimeInMillis()));
 		v.setLayoutParams(lp);
 		return v;
 	}
